+++ packages/eslint-plugin/src/rules/no-implied-eval.ts
@@ -99,6 +99,12 @@ export default util.createRule({
       return signatures.length > 0;
     }
 
+    function isBind(node: TSESTree.Node): boolean {
+      return node.type === AST_NODE_TYPES.MemberExpression
+        ? isBind(node.property)
+        : node.type === AST_NODE_TYPES.Identifier && node.name === 'bind';
+    }
+
     function isFunction(node: TSESTree.Node): boolean {
       switch (node.type) {
         case AST_NODE_TYPES.ArrowFunctionExpression:
@@ -106,20 +112,15 @@ export default util.createRule({
         case AST_NODE_TYPES.FunctionExpression:
           return true;
 
-        case AST_NODE_TYPES.MemberExpression:
-        case AST_NODE_TYPES.Identifier:
-        case AST_NODE_TYPES.ConditionalExpression:
-          return isFunctionType(node);
+        case AST_NODE_TYPES.Literal:
+        case AST_NODE_TYPES.TemplateLiteral:
+          return false;
 
         case AST_NODE_TYPES.CallExpression:
-          return (
-            (node.callee.type === AST_NODE_TYPES.Identifier &&
-              node.callee.name === 'bind') ||
-            isFunctionType(node)
-          );
+          return isBind(node.callee) || isFunctionType(node);
 
         default:
-          return false;
+          return isFunctionType(node);
       }
     }