+++ packages/typescript-estree/src/index.ts
@@ -1,4 +1,12 @@
-export * from './parser';
+export {
+  AST,
+  parse,
+  parseAndGenerateServices,
+  parseWithNodeMaps,
+  ParseAndGenerateServicesResult,
+  ParseWithNodeMapsResult,
+  clearProgramCache,
+} from './parser';
 export { ParserServices, TSESTreeOptions } from './parser-options';
 export { simpleTraverse } from './simple-traverse';
 export * from './ts-estree';
@@ -9,5 +17,5 @@ export { createProgramFromConfigFile as createProgram } from './create-program/u
 export { visitorKeys } from '@typescript-eslint/visitor-keys';
 
 // note - cannot migrate this to an import statement because it will make TSC copy the package.json to the dist folder
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access
 export const version: string = require('../package.json').version;